Dev: WIP (Organized some stuff)
